support downloading a whole file
Create one piece_validator task per piece and send valid pieces to the
piece_combiner task. piece_combiner writes each piece to the final file
by seeking to the position of the current piece.

The download_piece command is very inefficient because it downloads the
whole file and then extracts the requested piece.